Factor function-parameter utilities into a separate file
We are likely to grow more of these utilities, and they might
eventually be sunk down into a common library, so keep the factored
out.
