+++ Sources/SwiftSyntaxMacros/CMakeLists.txt
@@ -21,7 +21,9 @@ add_swift_host_library(SwiftSyntaxMacros
 
   AbstractSourceLocation.swift
   BasicMacroExpansionContext.swift
+  FunctionParameterUtils.swift
   MacroExpansionContext.swift
+  MacroReplacement.swift
   MacroSystem.swift
   Syntax+MacroEvaluation.swift
 )
